chore(bundle): phase 1 asset dedup and prerender split
