+++ index.html
@@ -14,6 +14,6 @@
   </head>
   <body>
     <div id="root"></div>
-    <script prerender type="module" src="/src/bootstrap.tsx"></script>
+    <script type="module" src="/src/bootstrap.tsx"></script>
   </body>
 </html>
